rew! opus rewrite
