+++ lib/modules/enrichment/merge-confidence/index.spec.ts
@@ -1,0 +1,602 @@
+import * as httpMock from '~test/http-mock.ts';
+import type { RenovateConfig } from '../../../config/types.ts';
+import { EXTERNAL_HOST_ERROR } from '../../../constants/error-messages.ts';
+import { logger } from '../../../logger/index.ts';
+import type { HostRule } from '../../../types/index.ts';
+import * as memCache from '../../../util/cache/memory/index.ts';
+import * as hostRules from '../../../util/host-rules.ts';
+import mergeConfidenceEnrichment, {
+  getMergeConfidenceLevel,
+  initConfig,
+  isActiveConfidenceLevel,
+  resetConfig,
+  satisfiesConfidenceLevel,
+} from './index.ts';
+
+describe('modules/enrichment/merge-confidence/index', () => {
+  const apiBaseUrl = 'https://www.baseurl.com/';
+  const defaultApiBaseUrl = 'https://developer.mend.io/';
+  const userSupportedDatasources = ['go'];
+
+  describe('isActiveConfidenceLevel()', () => {
+    it('returns false if null', () => {
+      expect(isActiveConfidenceLevel(null as never)).toBeFalse();
+    });
+
+    it('returns false if low', () => {
+      expect(isActiveConfidenceLevel('low')).toBeFalse();
+    });
+
+    it('returns false if nonsense', () => {
+      expect(isActiveConfidenceLevel('nonsense')).toBeFalse();
+    });
+
+    it('returns true if valid value (high)', () => {
+      expect(isActiveConfidenceLevel('high')).toBeTrue();
+    });
+  });
+
+  describe('satisfiesConfidenceLevel()', () => {
+    it('returns false if less', () => {
+      expect(satisfiesConfidenceLevel('low', 'high')).toBeFalse();
+    });
+
+    it('returns true if equal', () => {
+      expect(satisfiesConfidenceLevel('high', 'high')).toBeTrue();
+    });
+
+    it('returns true if more', () => {
+      expect(satisfiesConfidenceLevel('very high', 'high')).toBeTrue();
+    });
+  });
+
+  describe('API calling functions', () => {
+    const hostRule: HostRule = {
+      hostType: 'merge-confidence',
+      token: 'some-token',
+    };
+
+    beforeEach(() => {
+      hostRules.add(hostRule);
+      initConfig({ mergeConfidenceEndpoint: apiBaseUrl });
+      memCache.reset();
+    });
+
+    afterEach(() => {
+      hostRules.clear();
+      resetConfig();
+    });
+
+    describe('getMergeConfidenceLevel()', () => {
+      it('returns neutral if undefined updateType', async () => {
+        expect(
+          await getMergeConfidenceLevel(
+            'npm',
+            'renovate',
+            '25.0.0',
+            '25.0.0',
+            undefined as never,
+          ),
+        ).toBe('neutral');
+      });
+
+      it('returns neutral if irrelevant updateType', async () => {
+        expect(
+          await getMergeConfidenceLevel(
+            'npm',
+            'renovate',
+            '24.1.0',
+            '25.0.0',
+            'bump',
+          ),
+        ).toBe('neutral');
+      });
+
+      it('returns high if pinning', async () => {
+        expect(
+          await getMergeConfidenceLevel(
+            'npm',
+            'renovate',
+            '25.0.1',
+            '25.0.1',
+            'pin',
+          ),
+        ).toBe('high');
+      });
+
+      it('returns undefined if no token', async () => {
+        resetConfig();
+        hostRules.clear();
+
+        expect(
+          await getMergeConfidenceLevel(
+            'npm',
+            'renovate',
+            '24.2.0',
+            '25.0.0',
+            'major',
+          ),
+        ).toBeUndefined();
+      });
+
+      it('returns undefined if datasource is unsupported', async () => {
+        expect(
+          await getMergeConfidenceLevel(
+            'not-npm',
+            'renovate',
+            '24.2.0',
+            '25.0.0',
+            'major',
+          ),
+        ).toBeUndefined();
+      });
+
+      it('returns valid confidence level', async () => {
+        const datasource = 'npm';
+        const depName = 'renovate';
+        const currentVersion = '24.3.0';
+        const newVersion = '25.0.0';
+        httpMock
+          .scope(apiBaseUrl)
+          .get(
+            `/api/mc/json/${datasource}/${depName}/${currentVersion}/${newVersion}`,
+          )
+          .reply(200, { confidence: 'high' });
+
+        expect(
+          await getMergeConfidenceLevel(
+            datasource,
+            depName,
+            currentVersion,
+            newVersion,
+            'major',
+          ),
+        ).toBe('high');
+      });
+
+      it('escapes a package name containing a forward slash', async () => {
+        const datasource = 'npm';
+        const packageName = '@jest/global';
+        const escapedPackageName = '@jest%2fglobal';
+        const currentVersion = '24.3.0';
+        const newVersion = '25.0.0';
+        httpMock
+          .scope(apiBaseUrl)
+          .get(
+            `/api/mc/json/${datasource}/${escapedPackageName}/${currentVersion}/${newVersion}`,
+          )
+          .reply(200, { confidence: 'high' });
+
+        expect(
+          await getMergeConfidenceLevel(
+            datasource,
+            packageName,
+            currentVersion,
+            newVersion,
+            'major',
+          ),
+        ).toBe('high');
+      });
+
+      it('escapes a partial Maven coordinate of groupId:artifactId from the packageName', async () => {
+        const datasource = 'maven';
+        const packageName =
+          'org.springframework.boot:org.springframework.boot.gradle.plugin';
+        const escapedPackageName =
+          'org.springframework.boot%3Aorg.springframework.boot.gradle.plugin';
+        const currentVersion = '2.6.2';
+        const newVersion = '2.7.2';
+        httpMock
+          .scope(apiBaseUrl)
+          .get(
+            `/api/mc/json/${datasource}/${escapedPackageName}/${currentVersion}/${newVersion}`,
+          )
+          .reply(200, { confidence: 'high' });
+
+        expect(
+          await getMergeConfidenceLevel(
+            datasource,
+            packageName,
+            currentVersion,
+            newVersion,
+            'major',
+          ),
+        ).toBe('high');
+      });
+
+      it('returns neutral on invalid merge confidence response from api', async () => {
+        const datasource = 'npm';
+        const depName = 'renovate';
+        const currentVersion = '25.0.0';
+        const newVersion = '25.1.0';
+        httpMock
+          .scope(apiBaseUrl)
+          .get(
+            `/api/mc/json/${datasource}/${depName}/${currentVersion}/${newVersion}`,
+          )
+          .reply(200, { invalid: 'invalid' });
+
+        expect(
+          await getMergeConfidenceLevel(
+            datasource,
+            depName,
+            currentVersion,
+            newVersion,
+            'minor',
+          ),
+        ).toBe('neutral');
+      });
+
+      it('returns neutral on non 403/5xx error from API', async () => {
+        const datasource = 'npm';
+        const depName = 'renovate';
+        const currentVersion = '25.0.0';
+        const newVersion = '25.4.0';
+        httpMock
+          .scope(apiBaseUrl)
+          .get(
+            `/api/mc/json/${datasource}/${depName}/${currentVersion}/${newVersion}`,
+          )
+          .reply(400);
+
+        expect(
+          await getMergeConfidenceLevel(
+            datasource,
+            depName,
+            currentVersion,
+            newVersion,
+            'minor',
+          ),
+        ).toBe('neutral');
+
+        expect(logger.warn).toHaveBeenCalledWith(
+          expect.anything(),
+          'error fetching merge confidence data',
+        );
+      });
+
+      it('throws on 403-Forbidden response from API', async () => {
+        const datasource = 'npm';
+        const packageName = 'renovate';
+        const currentVersion = '25.0.0';
+        const newVersion = '25.4.0';
+        httpMock
+          .scope(apiBaseUrl)
+          .get(
+            `/api/mc/json/${datasource}/${packageName}/${currentVersion}/${newVersion}`,
+          )
+          .reply(403);
+
+        await expect(
+          getMergeConfidenceLevel(
+            datasource,
+            packageName,
+            currentVersion,
+            newVersion,
+            'minor',
+          ),
+        ).rejects.toThrow(EXTERNAL_HOST_ERROR);
+
+        expect(logger.error).toHaveBeenCalledWith(
+          expect.anything(),
+          'merge confidence API token rejected - aborting run',
+        );
+      });
+
+      it('throws on server error responses', async () => {
+        const datasource = 'npm';
+        const packageName = 'renovate';
+        const currentVersion = '25.0.0';
+        const newVersion = '25.4.0';
+        httpMock
+          .scope(apiBaseUrl)
+          .get(
+            `/api/mc/json/${datasource}/${packageName}/${currentVersion}/${newVersion}`,
+          )
+          .reply(503);
+
+        await expect(
+          getMergeConfidenceLevel(
+            datasource,
+            packageName,
+            currentVersion,
+            newVersion,
+            'minor',
+          ),
+        ).rejects.toThrow(EXTERNAL_HOST_ERROR);
+
+        expect(logger.error).toHaveBeenCalledWith(
+          expect.anything(),
+          'merge confidence API failure: 5xx - aborting run',
+        );
+      });
+
+      it('returns high if pinning digest', async () => {
+        expect(
+          await getMergeConfidenceLevel(
+            'npm',
+            'renovate',
+            '25.0.1',
+            '25.0.1',
+            'pinDigest',
+          ),
+        ).toBe('high');
+      });
+    });
+
+    describe('EnrichmentApi', () => {
+      describe('isEnabled()', () => {
+        it('returns true when token and apiBaseUrl are set', () => {
+          expect(
+            mergeConfidenceEnrichment.isEnabled({} as RenovateConfig),
+          ).toBeTrue();
+        });
+
+        it('returns false when not configured', () => {
+          resetConfig();
+          hostRules.clear();
+          initConfig({});
+          expect(
+            mergeConfidenceEnrichment.isEnabled({} as RenovateConfig),
+          ).toBeFalse();
+        });
+      });
+
+      describe('init()', () => {
+        beforeEach(() => {
+          resetConfig();
+        });
+
+        it('using default base url and supported datasources if either is set', async () => {
+          httpMock
+            .scope(defaultApiBaseUrl)
+            .get(`/api/mc/availability`)
+            .reply(200);
+
+          await expect(
+            mergeConfidenceEnrichment.init!({} as RenovateConfig),
+          ).toResolve();
+
+          expect(logger.debug).toHaveBeenCalledWith(
+            {
+              supportedDatasources: [
+                'go',
+                'maven',
+                'npm',
+                'nuget',
+                'packagist',
+                'pypi',
+                'rubygems',
+              ],
+            },
+            'merge confidence API - successfully authenticated',
+          );
+        });
+
+        it('warns and then resolves if base url is invalid', async () => {
+          httpMock
+            .scope(defaultApiBaseUrl)
+            .get(`/api/mc/availability`)
+            .reply(200);
+
+          await expect(
+            mergeConfidenceEnrichment.init!({
+              mergeConfidenceEndpoint: 'invalid-url.com',
+            } as RenovateConfig),
+          ).toResolve();
+
+          expect(logger.warn).toHaveBeenCalledWith(
+            expect.anything(),
+            'invalid merge confidence API base URL found in environment variables - using default value instead',
+          );
+
+          expect(logger.debug).toHaveBeenCalledWith(
+            expect.anything(),
+            'merge confidence API - successfully authenticated',
+          );
+        });
+
+        it('uses custom supported datasources and a base URL containing a path', async () => {
+          const renovateApi = 'https://domain.com/proxy/renovate-api';
+          httpMock.scope(renovateApi).get(`/api/mc/availability`).reply(200);
+
+          await expect(
+            mergeConfidenceEnrichment.init!({
+              mergeConfidenceEndpoint: renovateApi,
+              mergeConfidenceDatasources: userSupportedDatasources,
+            } as RenovateConfig),
+          ).toResolve();
+
+          expect(logger.trace).toHaveBeenCalledWith(
+            expect.anything(),
+            'using merge confidence API base found in environment variables',
+          );
+
+          expect(logger.debug).toHaveBeenCalledWith(
+            {
+              supportedDatasources: ['go'],
+            },
+            'merge confidence API - successfully authenticated',
+          );
+        });
+
+        it('resolves if no token', async () => {
+          hostRules.clear();
+
+          await expect(
+            mergeConfidenceEnrichment.init!({} as RenovateConfig),
+          ).toResolve();
+
+          expect(logger.trace).toHaveBeenCalledWith(
+            'merge confidence API usage is disabled',
+          );
+        });
+
+        it('resolves when token is valid', async () => {
+          httpMock.scope(apiBaseUrl).get(`/api/mc/availability`).reply(200);
+
+          await expect(
+            mergeConfidenceEnrichment.init!({
+              mergeConfidenceEndpoint: apiBaseUrl,
+            } as RenovateConfig),
+          ).toResolve();
+
+          expect(logger.debug).toHaveBeenCalledWith(
+            expect.anything(),
+            'merge confidence API - successfully authenticated',
+          );
+        });
+
+        it('throws on 403-Forbidden from mc API', async () => {
+          httpMock.scope(apiBaseUrl).get(`/api/mc/availability`).reply(403);
+
+          await expect(
+            mergeConfidenceEnrichment.init!({
+              mergeConfidenceEndpoint: apiBaseUrl,
+            } as RenovateConfig),
+          ).rejects.toThrow(EXTERNAL_HOST_ERROR);
+
+          expect(logger.error).toHaveBeenCalledWith(
+            expect.anything(),
+            'merge confidence API token rejected - aborting run',
+          );
+        });
+
+        it('throws on 5xx host errors from mc API', async () => {
+          httpMock.scope(apiBaseUrl).get(`/api/mc/availability`).reply(503);
+
+          await expect(
+            mergeConfidenceEnrichment.init!({
+              mergeConfidenceEndpoint: apiBaseUrl,
+            } as RenovateConfig),
+          ).rejects.toThrow(EXTERNAL_HOST_ERROR);
+
+          expect(logger.error).toHaveBeenCalledWith(
+            expect.anything(),
+            'merge confidence API failure: 5xx - aborting run',
+          );
+        });
+
+        it('throws on ECONNRESET', async () => {
+          httpMock
+            .scope(apiBaseUrl)
+            .get(`/api/mc/availability`)
+            .replyWithError(httpMock.error({ code: 'ECONNRESET' }));
+
+          await expect(
+            mergeConfidenceEnrichment.init!({
+              mergeConfidenceEndpoint: apiBaseUrl,
+            } as RenovateConfig),
+          ).rejects.toThrow(EXTERNAL_HOST_ERROR);
+
+          expect(logger.error).toHaveBeenCalledWith(
+            expect.anything(),
+            'merge confidence API request failed - aborting run',
+          );
+        });
+      });
+
+      describe('enrichUpdate()', () => {
+        it('returns mergeConfidenceLevel for a supported update', async () => {
+          const datasource = 'npm';
+          const packageName = 'renovate';
+          const currentVersion = '24.3.0';
+          const newVersion = '25.0.0';
+          httpMock
+            .scope(apiBaseUrl)
+            .get(
+              `/api/mc/json/${datasource}/${packageName}/${currentVersion}/${newVersion}`,
+            )
+            .reply(200, { confidence: 'high' });
+
+          const result = await mergeConfidenceEnrichment.enrichUpdate!(
+            {
+              datasource,
+              packageName,
+              currentVersion,
+              newVersion,
+              updateType: 'major',
+            },
+            {} as RenovateConfig,
+          );
+
+          expect(result).toEqual({ mergeConfidenceLevel: 'high' });
+        });
+
+        it('returns null when no token is configured', async () => {
+          resetConfig();
+          hostRules.clear();
+          initConfig({});
+
+          const result = await mergeConfidenceEnrichment.enrichUpdate!(
+            {
+              datasource: 'npm',
+              packageName: 'renovate',
+              currentVersion: '24.0.0',
+              newVersion: '25.0.0',
+              updateType: 'major',
+            },
+            {} as RenovateConfig,
+          );
+
+          expect(result).toBeNull();
+        });
+
+        it('returns null when currentVersion is missing', async () => {
+          const result = await mergeConfidenceEnrichment.enrichUpdate!(
+            {
+              datasource: 'npm',
+              packageName: 'renovate',
+              newVersion: '25.0.0',
+              updateType: 'major',
+            },
+            {} as RenovateConfig,
+          );
+
+          expect(result).toBeNull();
+        });
+
+        it('returns null when updateType is missing', async () => {
+          const result = await mergeConfidenceEnrichment.enrichUpdate!(
+            {
+              datasource: 'npm',
+              packageName: 'renovate',
+              currentVersion: '24.0.0',
+              newVersion: '25.0.0',
+            },
+            {} as RenovateConfig,
+          );
+
+          expect(result).toBeNull();
+        });
+
+        it('returns mapped confidence for pin updateType', async () => {
+          const result = await mergeConfidenceEnrichment.enrichUpdate!(
+            {
+              datasource: 'npm',
+              packageName: 'renovate',
+              currentVersion: '25.0.1',
+              newVersion: '25.0.1',
+              updateType: 'pin',
+            },
+            {} as RenovateConfig,
+          );
+
+          expect(result).toEqual({ mergeConfidenceLevel: 'high' });
+        });
+      });
+
+      describe('cleanup()', () => {
+        it('resets config', async () => {
+          expect(
+            mergeConfidenceEnrichment.isEnabled({} as RenovateConfig),
+          ).toBeTrue();
+          await mergeConfidenceEnrichment.cleanup!();
+          expect(
+            mergeConfidenceEnrichment.isEnabled({} as RenovateConfig),
+          ).toBeFalse();
+        });
+      });
+    });
+  });
+});
