--- lib/util/merge-confidence/types.ts
@@ -1,3 +1,0 @@
-import type { MERGE_CONFIDENCE } from './common.ts';
-
-export type MergeConfidence = (typeof MERGE_CONFIDENCE)[number];
