+++ lib/workers/global/initialize.ts
@@ -12,7 +12,6 @@ import { setEmojiConfig } from '../../util/emoji.ts';
 import { validateGitVersion } from '../../util/git/index.ts';
 import * as hostRules from '../../util/host-rules.ts';
 import { setHttpRateLimits } from '../../util/http/rate-limits.ts';
-import { initMergeConfidence } from '../../util/merge-confidence/index.ts';
 import { setMaxLimit } from './limits.ts';
 
 async function setDirectories(input: AllConfig): Promise<AllConfig> {
@@ -94,7 +93,6 @@ export async function globalInitialize(
   setEmojiConfig(config);
   setGlobalHostRules(config);
   configureThirdPartyLibraries(config);
-  await initMergeConfidence(config);
   await initEnrichments(config);
   return config;
 }
